перенос получения треков на страницу minePage
