+++ src/App.css
@@ -73,8 +73,9 @@ body {
 }
 
 .container {
-  max-width: 1920px;
-  max-height: 1000px;
+  max-width: 100%;
+  height: 100%;
+  /* max-height: 1000px; */
   /* height: 100vh; */
   margin: 0 auto;
   position: relative;
