+++ src/components/PlayListContent/playListContent.styled.js
@@ -1,17 +1,4 @@
 import  styled  from 'styled-components';
-// const blinkAnimation = keyframes`
-//     0% {
-//     opacity: 0;
-//   }
-//   50% {
-//     opacity: 1;
-//   }
-//   100% {
-//     opacity: 0;
-//   }
-//   `;
-  
-
 
 export const PlaylistItem = styled.div`
     width: 100%;
