+++ src/store/actions/types/index.js
@@ -2,3 +2,4 @@ export const SET_CURRENT_TRACK = "SET_CURRENT_TRACK";
 export const NEXT_TRACK = "NEXT_TRACK";
 export const PREV_TRACK = "PREV_TRACK";
 export const TOGGLE_SUFFLED = "TOGGLE_SUFFLED";
+export const ALL_TRACKS_LOADING = "ALL_TRACKS_LOADING"
